# leaderboard/public/openapi.json
{"openapi":"3.0.0","info":{"title":"Leaderboard API","version":"","description":"Leaderboard api documentation.","termsOfService":"","license":null,"contact":[],"x-logo":{"url":"","altText":"Leaderboard API","backgroundColor":""}},"components":{"securitySchemes":{"BearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}},"schemas":[]},"servers":[{"url":"http:\/\/127.0.0.1:8000","description":"Local server."}],"paths":{"\/api\/admin\/users":{"post":{"tags":["Authorization and user management"],"operationId":"Create a user","description":"Enpoint to create a api admin user","requestBody":{"description":"Enpoint to create a api admin user","required":true,"content":{"application\/json":{"schema":{"type":"object","required":["name","email","password"],"properties":{"name":{"type":"string","example":"c!\"Qc'KD#Q[V","description":"The username"},"email":{"type":"email","example":";Eg[ltA^bSy~","description":"The users email"},"password":{"type":"string","example":"B$yKsq0Mp(WP","description":"The admins user's"}},"example":{"name":"c!\"Qc'KD#Q[V","email":";Eg[ltA^bSy~","password":"B$yKsq0Mp(WP"}}}}},"parameters":[],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"name":"Username","email":"Email"}}}}},"400":{"description":"error","content":{"application\/json":{"schema":{"type":"object","example":{"message":"Invalid request"}}}}},"409":{"description":"error","content":{"application\/json":{"schema":{"type":"object","example":{"message":"User email already exist"}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X POST \"http:\/\/localhost\/api\/admin\/users\" \n    -d \"name\"=\"c!\"Qc'KD#Q[V\" \\\n    -d \"email\"=\";Eg[ltA^bSy~\" \\\n    -d \"password\"=\"B$yKsq0Mp(WP\" "},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost\/api\/admin\/users\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nlet body = JSON.stringify({\n    \"name\": \"c!&quot;Qc&#039;KD#Q[V\",\n    \"email\": \";Eg[ltA^bSy~\",\n    \"password\": \"B$yKsq0Mp(WP\",\n})\n\nfetch(url, {\n    method: \"POST\",\n    headers: headers,\n    body: body\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n$body = [\n    \"name\" =>  \"c!&quot;Qc&#039;KD#Q[V\",\n    \"email\" =>  \";Eg[ltA^bSy~\",\n    \"password\" =>  \"B$yKsq0Mp(WP\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost\/api\/admin\/users\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"POST\");\ncurl_setopt($ch, CURLOPT_POSTFIELDS, $body);\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}},"\/api\/admin\/users\/token":{"post":{"tags":["Authorization and user management"],"operationId":"Create authorize token","description":"Enpoint to create a BearerAuth token","requestBody":{"description":"Enpoint to create a BearerAuth token","required":true,"content":{"application\/json":{"schema":{"type":"object","required":["email","password"],"properties":{"email":{"type":"email","example":"ECcycu_'G*O+","description":"The users email"},"password":{"type":"string","example":"JpC2#On&}6\/(","description":"The admins user's password"}},"example":{"email":"ECcycu_'G*O+","password":"JpC2#On&}6\/("}}}}},"parameters":[],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{"token":"BearerAuthToken"}}}}},"403":{"description":"error","content":{"application\/json":{"schema":{"type":"object","example":{"message":"Unauthorized"}}}}},"404":{"description":"error","content":{"application\/json":{"schema":{"type":"object","example":{"message":"User not found"}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X POST \"http:\/\/localhost\/api\/admin\/users\/token\" \n    -d \"email\"=\"ECcycu_'G*O+\" \\\n    -d \"password\"=\"JpC2#On&}6\/(\" "},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost\/api\/admin\/users\/token\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nlet body = JSON.stringify({\n    \"email\": \"ECcycu_&#039;G*O+\",\n    \"password\": \"JpC2#On&amp;}6\/(\",\n})\n\nfetch(url, {\n    method: \"POST\",\n    headers: headers,\n    body: body\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n$body = [\n    \"email\" =>  \"ECcycu_&#039;G*O+\",\n    \"password\" =>  \"JpC2#On&amp;}6\/(\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost\/api\/admin\/users\/token\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"POST\");\ncurl_setopt($ch, CURLOPT_POSTFIELDS, $body);\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}},"\/api\/leaderboards":{"post":{"tags":["Leaderboard admin"],"operationId":"","description":"","requestBody":{"description":"","required":true,"content":{"application\/json":{"schema":{"type":"object"}}}},"parameters":[],"responses":{"401":{"description":"error","content":{"application\/json":{"schema":{"type":"object","example":{"message":"Unauthenticated."}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X POST \"http:\/\/localhost\/api\/leaderboards\" \n"},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost\/api\/leaderboards\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"POST\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost\/api\/leaderboards\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"POST\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]},"get":{"tags":["Leaderboard admin"],"operationId":"Get leaderboards list","description":"Enpoint get the leaderboard list","parameters":[],"responses":{"200":{"description":"success","content":{"application\/json":{"schema":{"type":"object","example":{}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X GET -G \"http:\/\/localhost\/api\/leaderboards\" \n"},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost\/api\/leaderboards\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"GET\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost\/api\/leaderboards\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}},"\/api\/leaderboards\/{leaderboard_id}":{"delete":{"tags":["Leaderboard admin"],"operationId":"This method performs a soft delete in order to avoid data lost by mistake","description":"","parameters":[],"responses":{"401":{"description":"error","content":{"application\/json":{"schema":{"type":"object","example":{"message":"Unauthenticated."}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X DELETE \"http:\/\/localhost\/api\/leaderboards\/{leaderboard_id}\" \n"},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost\/api\/leaderboards\/{leaderboard_id}\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"DELETE\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost\/api\/leaderboards\/{leaderboard_id}\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"DELETE\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}},"\/api\/leaderboards\/{leaderboard_id}\/enable":{"patch":{"tags":["Leaderboard admin"],"operationId":"","description":"","requestBody":{"description":"","required":true,"content":{"application\/json":{"schema":{"type":"object"}}}},"parameters":[],"responses":{"401":{"description":"error","content":{"application\/json":{"schema":{"type":"object","example":{"message":"Unauthenticated."}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X PATCH \"http:\/\/localhost\/api\/leaderboards\/{leaderboard_id}\/enable\" \n"},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost\/api\/leaderboards\/{leaderboard_id}\/enable\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"PATCH\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost\/api\/leaderboards\/{leaderboard_id}\/enable\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"PATCH\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}},"\/api\/leaderboards\/{leaderboard_id}\/players":{"post":{"tags":["general"],"operationId":"","description":"","requestBody":{"description":"","required":true,"content":{"application\/json":{"schema":{"type":"object"}}}},"parameters":[],"responses":{"401":{"description":"error","content":{"application\/json":{"schema":{"type":"object","example":{"message":"Unauthenticated."}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X POST \"http:\/\/localhost\/api\/leaderboards\/{leaderboard_id}\/players\" \n"},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost\/api\/leaderboards\/{leaderboard_id}\/players\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"POST\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost\/api\/leaderboards\/{leaderboard_id}\/players\");\ncurl_setopt($ch, CURLOPT_CUSTOMREQUEST, \"POST\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]},"get":{"tags":["general"],"operationId":"","description":"","parameters":[],"responses":{"401":{"description":"error","content":{"application\/json":{"schema":{"type":"object","example":{"message":"Unauthenticated."}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X GET -G \"http:\/\/localhost\/api\/leaderboards\/{leaderboard_id}\/players\" \n"},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost\/api\/leaderboards\/{leaderboard_id}\/players\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"GET\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost\/api\/leaderboards\/{leaderboard_id}\/players\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}},"\/api\/leaderboards\/{leaderboard_id}\/players\/{player_id}\/rank":{"get":{"tags":["general"],"operationId":"","description":"","parameters":[],"responses":{"401":{"description":"error","content":{"application\/json":{"schema":{"type":"object","example":{"message":"Unauthenticated."}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X GET -G \"http:\/\/localhost\/api\/leaderboards\/{leaderboard_id}\/players\/{player_id}\/rank\" \n"},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost\/api\/leaderboards\/{leaderboard_id}\/players\/{player_id}\/rank\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"GET\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost\/api\/leaderboards\/{leaderboard_id}\/players\/{player_id}\/rank\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}},"\/api\/leaderboards\/{leaderboard_id}\/players\/{player_id}\/rank\/around\/{records_around}":{"get":{"tags":["general"],"operationId":"","description":"","parameters":[],"responses":{"401":{"description":"error","content":{"application\/json":{"schema":{"type":"object","example":{"message":"Unauthenticated."}}}}}},"x-code-samples":[{"lang":"Bash","source":"curl -X GET -G \"http:\/\/localhost\/api\/leaderboards\/{leaderboard_id}\/players\/{player_id}\/rank\/around\/{records_around}\" \n"},{"lang":"Javascript","source":"const url = new URL(\"http:\/\/localhost\/api\/leaderboards\/{leaderboard_id}\/players\/{player_id}\/rank\/around\/{records_around}\");\n\nlet headers = {\n    \"Accept\": \"application\/json\",\n    \"Content-Type\": \"application\/json\",\n}\n\nfetch(url, {\n    method: \"GET\",\n    headers: headers,\n})\n    .then(response => response.json())\n    .then(json => console.log(json));\n"},{"lang":"PHP","source":"$headers = [\n    \"Accept\" => \"application\/json\",\n    \"Content-Type\" => \"application\/json\",\n];\n\n\n$ch = curl_init(\"http:\/\/localhost\/api\/leaderboards\/{leaderboard_id}\/players\/{player_id}\/rank\/around\/{records_around}\");\ncurl_setopt($ch, CURLOPT_HTTPHEADER, $headers);\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$result = curl_exec($ch);\n$err = curl_error($ch);\ncurl_close($ch);\n\nif ($err) {\n  throw \\Exception($err);\n} else {\n  return $response;\n}\n"}]}}},"x-tagGroups":[]}